修复 ResizablePane 组件中的精度丢失问题，确保控制器值转换为双精度浮点数。同时更新 ColorSchemes 类中的侧边栏颜色配置，优化视觉效果。

+++ lib/src/components/layout/resizable.dart
@@ -380,7 +380,8 @@ class _ResizablePaneState extends State<ResizablePane> {
           if (controller.collapsed) {
             size = widget.collapsedSize;
           } else {
-            size = controller.value;
+            // 防止精度丢失
+            size = controller.value.toInt().toDouble();
           }
         } else if (_controller is FlexibleResizablePaneController) {
           final controller = _controller as FlexibleResizablePaneController;
